@@ -3,14 +3,13 @@ description: Demonstrates how to use the month_picker_dialog_2 plugin.
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.17.1 <3.0.0'
+  sdk: '>=2.19.2 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  cupertino_icons: ^1.0.2  
 
 dev_dependencies:
   flutter_test:
