@@ -1,10 +1,10 @@
 name: month_picker_dialog_2
 description: Internationalized dialog for picking a single month from an infinite list of years.
-version: 0.6.1+2
+version: 0.6.2
 homepage: https://github.com/Macacoazul01/month_picker_dialog_2
 
 environment:
-  sdk: '>=2.17.1 <3.0.0'
+  sdk: '>=2.19.2 <3.0.0'
 
 dependencies:
   flutter:
