@@ -54,7 +54,7 @@ class _MyAppState extends State<MyApp> {
         body: Center(
           child: Text(
             'Year: ${selectedDate?.year}\nMonth: ${selectedDate?.month}',
-            style: Theme.of(context).textTheme.headline4,
+            style: Theme.of(context).textTheme.headlineMedium,
             textAlign: TextAlign.center,
           ),
         ),
